Add endpoints for filtering consultations by test statuses
Introduced three new endpoints in `ClinicController` to retrieve consultations based on prescription, lab test, and imaging test statuses. Updated service and repository layers to handle these operations, including SQL queries for fetching appropriate data. Modified `TokenRequest` to support `Short` type for status handling.

+++ src/main/kotlin/com/icare/model/TokenRequest.kt
@@ -1,3 +1,3 @@
 package com.icare.model
 
-data class TokenRequest(val token: String,val status:Int=0)
+data class TokenRequest(val token: String,val status: Short=0)
